fix hero struct
공격속도 타입 변환 및 애니메이션에 공격 속도로 Play Rate 추가

+++ Source/Porter/Hero/PHero.h
@@ -45,7 +45,7 @@ protected:
 	int Damage;
 
 	UPROPERTY(EditAnywhere, Category = Stat, Meta = (AllowPrivateAccess = "true"))
-	int AttackSpeed;
+	float AttackSpeed;
 
 	UPROPERTY(EditAnywhere, Category = Mesh, Meta = (AllowPrivateAccess = "true"))
 	FVector MeshLocation;
